+++ packages/cypress/src/generators/cypress-project/files/v10-and-after/cypress.config.ts__tmpl__
@@ -1,17 +1,6 @@
 import { defineConfig } from 'cypress';
+import { nxE2EPreset } from '@nrwl/cypress/plugins/cypress-preset';
 
 export default defineConfig({
-  e2e: {
-    fileServerFolder: '.',
-    fixturesFolder: 'src/fixtures',
-    integrationFolder: 'src/e2e',
-    supportFile: 'src/support/e2e.<%= ext %>',
-    specPattern: 'src/**/*.cy.{js,ts}',
-    pluginsFile: false,
-    video: true,
-    videosFolder: '<%= offsetFromRoot %>dist/cypress/<%= projectRoot %>/videos',
-    screenshotsFolder:
-      '<%= offsetFromRoot %>dist/cypress/<%= projectRoot %>/screenshots',
-    chromeWebSecurity: false,
-  },
+  e2e: nxE2EPreset(__dirname)
 });
