+++ packages/web/src/utils/web.config.ts
@@ -276,7 +276,12 @@ export function getPolyfillsPartial(
   return config;
 }
 
-function getCSSModuleLocalIdent(context, localIdentName, localName, options) {
+export function getCSSModuleLocalIdent(
+  context,
+  localIdentName,
+  localName,
+  options
+) {
   // Use the filename or folder name, based on some uses the index.js / index.module.(css|scss|sass) project style
   const fileNameOrFolder = context.resourcePath.match(
     /index\.module\.(css|scss|sass|styl)$/
