+++ packages/workspace/src/utils/rules/rename-npm-packages.spec.ts
@@ -1,14 +1,14 @@
 import { Tree } from '@angular-devkit/schematics';
 import { UnitTestTree } from '@angular-devkit/schematics/testing';
-import { readJsonInTree } from '../ast-utils';
+import { createEmptyWorkspace } from '../../../testing';
 import {
   callRule,
-  runSchematic,
   runExternalSchematic,
+  runSchematic,
 } from '../../utils/testing';
-import { createEmptyWorkspace } from '../../../testing';
+import { readJsonInTree } from '../ast-utils';
 
-import { renameNpmPackages, PackageRenameMapping } from './rename-npm-packages';
+import { renameNpmPackages } from './rename-npm-packages';
 
 describe('renameNpmPackages Rule', () => {
   let tree: UnitTestTree;
@@ -93,6 +93,7 @@ describe('renameNpmPackages Rule', () => {
 
     tree = await runSchematic('lib', { name: 'library-1' }, tree);
     tree = await runSchematic('lib', { name: 'library-2' }, tree);
+    // dies right here 🤔
     tree = await runExternalSchematic(
       '@nrwl/angular',
       'application',
